nwjs master

+++ deps/uv/src/unix/getaddrinfo.c
@@ -90,7 +90,7 @@ int uv__getaddrinfo_translate_error(int sys_err) {
   case EAI_SYSTEM: return UV__ERR(errno);
 #endif
   }
-  assert(!"unknown EAI_* error code");
+  assert(!(int)("unknown EAI_* error code"));
   abort();
   return 0;  /* Pacify compiler. */
 }
